prune options working
