+++ src/vorta/views/prune_options_page.py
@@ -1,0 +1,52 @@
+from PyQt6 import uic
+
+from vorta.store.models import BackupProfileMixin
+from vorta.utils import format_archive_name, get_asset
+
+uifile = get_asset('UI/prune_options_page.ui')
+PruneOptionsUI, PruneOptionsBase = uic.loadUiType(uifile)
+
+
+class PruneOptionsPage(PruneOptionsBase, PruneOptionsUI, BackupProfileMixin):
+    prune_intervals = ['hour', 'day', 'week', 'month', 'year']
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setupUi(self)
+        self.populate_from_profile()
+        self.archiveNameTemplate.textChanged.connect(
+            lambda tpl, key='new_archive_name': self.save_archive_template(tpl, key)
+        )
+        self.prunePrefixTemplate.textChanged.connect(
+            lambda tpl, key='prune_prefix': self.save_archive_template(tpl, key)
+        )
+
+    def populate_from_profile(self):
+        # Populate pruning options from database
+        profile = self.profile()
+        for i in self.prune_intervals:
+            getattr(self, f'prune_{i}').setValue(getattr(profile, f'prune_{i}'))
+            getattr(self, f'prune_{i}').valueChanged.connect(self.save_prune_setting)
+        self.prune_keep_within.setText(profile.prune_keep_within)
+        self.prune_keep_within.editingFinished.connect(self.save_prune_setting)
+
+    def save_prune_setting(self, new_value=None):
+        profile = self.profile()
+        for i in self.prune_intervals:
+            setattr(profile, f'prune_{i}', getattr(self, f'prune_{i}').value())
+        profile.prune_keep_within = self.prune_keep_within.text()
+        profile.save()
+
+    def save_archive_template(self, tpl, key):
+        profile = self.profile()
+        try:
+            preview = self.tr('Preview: %s') % format_archive_name(profile, tpl)
+            setattr(profile, key, tpl)
+            profile.save()
+        except Exception:
+            preview = self.tr('Error in archive name template.')
+
+        if key == 'new_archive_name':
+            self.archiveNamePreview.setText(preview)
+        else:
+            self.prunePrefixPreview.setText(preview)
